% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discard_non_replicated_samples.R
\name{discard_non_replicated_samples}
\alias{discard_non_replicated_samples}
\title{Get rid of all samples that have not enough biological replicates}
\usage{
discard_non_replicated_samples(df, min_replicates = 3)
}
\value{
A dataframe object
}
\description{
Get rid of all samples that have not enough biological replicates
}
