% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_label_language.R
\name{reduce_label_language}
\alias{reduce_label_language}
\title{Simplify labels of your categorical variables}
\usage{
reduce_label_language(df)
}
\value{
A dataframe object
}
\description{
Get rid of all unnecessarily complexities in your dataframe labels in order to simplify the language
}
